requirements.txt
boto3
ansicolors